fixed machine learning
