+++ api/app/tests/common.py
@@ -206,3 +206,10 @@ def default_mock_requests_session_post(self, url, data=None, json=None, **kwargs
     """ Return a mocked request response from a request.Session object """
     return default_mock_requests_post(url, data, json, **kwargs)
 # pylint: enable=redefined-outer-name
+
+
+def str2float(value: str):
+    """ Change a string into a floating point number, or a None """
+    if value == 'None':
+        return None
+    return float(value)
