+++ api/app/tests/fba_calc/test_fba.py
@@ -9,15 +9,14 @@ Running the test with this syntax `pytest app/tests/fba/test_fba.py -s` will res
 JSON to be printed out to screen. If the tests are failing, a developer can inspect the JSON and establish
 if the change in output is due to a bug, or due to a valid change.
 """
-import json
 from typing import Tuple
-from pytest_bdd import scenario, given, then, parsers
+from pytest_bdd import scenario, given, parsers
 from fastapi.testclient import TestClient
 from aiohttp import ClientSession
 import pytest
 import app.main
 from app.tests.common import default_mock_client_get
-from app.tests import load_json_file, load_json_file_with_name
+from app.tests import load_json_file_with_name
 
 
 @pytest.mark.usefixtures('mock_jwt_decode')
@@ -46,13 +45,3 @@ def given_request(monkeypatch, request_json: Tuple[dict, str]):
         'response': client.post('/api/fba-calc/stations', headers=headers, json=request_json[0]),
         'filename': request_json[1]
     }
-
-
-@then(parsers.parse("the response is {response_json}"),
-      converters={'response_json': load_json_file(__file__)})
-def then_response(result, response_json: dict):
-    """ Check entire response """
-    if response_json is not None:
-        print('actual:\n{}'.format(json.dumps(result['response'].json(), indent=4)))
-        print('expected:\n{}'.format(json.dumps(response_json, indent=4)))
-        assert result['response'].json() == response_json, result['filename']
