+++ api/app/tests/fba_calc/test_fba_error.py
@@ -9,7 +9,8 @@ from app.schemas.fba_calc import FuelTypeEnum
 from app.utils.time import get_hour_20_from_date
 from app.fba_calculator import calculate_fire_behaviour_advisory, FBACalculatorWeatherStation
 from app.utils.redapp import FBPCalculateStatisticsCOM
-from app.tests.fba_calc import (str2float, check_metric, acceptable_margin_of_error,
+from app.tests.common import str2float
+from app.tests.fba_calc import (check_metric, acceptable_margin_of_error,
                                 fire_size_acceptable_margin_of_error)
 import pytest
