+++ api/app/tests/fba_calc/test_fba_error_random_sample.py
@@ -2,6 +2,7 @@
 Unit tests for fire behavour calculator.
 """
 from datetime import datetime, timezone as dt_tz
+from multiprocessing.spawn import import_main_path
 import random
 from typing import Final
 import logging
@@ -12,7 +13,8 @@ from app.utils.time import get_hour_20_from_date
 from app.fba_calculator import calculate_fire_behaviour_advisory, FBACalculatorWeatherStation
 from app.utils.redapp import FBPCalculateStatisticsCOM
 from app.utils.cffdrs import initial_spread_index, bui_calc
-from app.tests.fba_calc import str2float, check_metric, fire_size_acceptable_margin_of_error
+from app.tests.fba_calc import check_metric, fire_size_acceptable_margin_of_error
+from app.tests.common import str2float
 import pytest
