+++ api/app/tests/test_fba_endpoint.py
@@ -29,12 +29,6 @@ def given_fba_fire_centers_request(monkeypatch):
     return client.get('/api/fba/fire-centers/', headers=headers)
 
 
-@then(parsers.parse('the response status code is {status}'), converters={'status': int})
-def assert_status_code(response, status):
-    """ Assert that we receive the expected status code """
-    assert response.status_code == status
-
-
 @then(parsers.parse('the response contains the list of {expected_fire_centers}'), converters={'expected_fire_centers': load_json_file(__file__)})
 def assert_fire_centers_list(response, expected_fire_centers):
     """ Assert that the fire centers returned are what is from the middleware """
