+++ api/app/tests/weather_models/endpoints/test_models_endpoints.feature
@@ -7,9 +7,9 @@ Feature: Generic model endpoint testing
         Then The status code = <expected_status_code>
         Then The response = <expected_response>
         Examples:
-            | codes      | endpoint                                          | crud_mapping                                                           | expected_status_code | expected_response                                                 | notes                                                                                                            |
-            | [322]      | /api/weather_models/GDPS/predictions/summaries/   | test_models_predictions_summaries_crud_mappings.json                   | 200                  | test_models_predictions_summaries_response.json                   | n/a                                                                                                              |
-            | [322, 838] | /api/weather_models/GDPS/predictions/summaries/   | test_models_predictions_summaries_multiple_crud_mappings.json          | 200                  | test_models_predictions_summaries_response_multiple.json          | n/a                                                                                                              |
-            | [838]      | /api/weather_models/GDPS/predictions/most_recent/ | test_models_predictions_most_recent_GDPS_[838]_crud_mappings.json      | 200                  | test_models_predictions_most_recent_GDPS_[838]_response.json      | n/a                                                                                                              |
-            | [838, 209] | /api/weather_models/GDPS/predictions/most_recent/ | test_models_predictions_most_recent_RDPS_[838, 209]_crud_mappings.json | 200                  | test_models_predictions_most_recent_RDPS_[838, 209]_response.json | n/a                                                                                                              |
-            | [956]      | /api/weather_models/GDPS/predictions/most_recent/ | test_models_predictions_most_recent_GDPS_[956]_crud_mappings.json      | 200                  | test_models_predictions_most_recent_GDPS_[956]_response.json      | Testing for rare case, where model run time matches prediction time, and previous precip delta is not available. |
+            | codes      | endpoint                                          | crud_mapping                                                                                    | expected_status_code | expected_response                                                 | notes                                                                                                            |
+            | [322]      | /api/weather_models/GDPS/predictions/summaries/   | weather_models/endpoints/test_models_predictions_summaries_crud_mappings.json                   | 200                  | test_models_predictions_summaries_response.json                   | n/a                                                                                                              |
+            | [322, 838] | /api/weather_models/GDPS/predictions/summaries/   | weather_models/endpoints/test_models_predictions_summaries_multiple_crud_mappings.json          | 200                  | test_models_predictions_summaries_response_multiple.json          | n/a                                                                                                              |
+            | [838]      | /api/weather_models/GDPS/predictions/most_recent/ | weather_models/endpoints/test_models_predictions_most_recent_GDPS_[838]_crud_mappings.json      | 200                  | test_models_predictions_most_recent_GDPS_[838]_response.json      | n/a                                                                                                              |
+            | [838, 209] | /api/weather_models/GDPS/predictions/most_recent/ | weather_models/endpoints/test_models_predictions_most_recent_RDPS_[838, 209]_crud_mappings.json | 200                  | test_models_predictions_most_recent_RDPS_[838, 209]_response.json | n/a                                                                                                              |
+            | [956]      | /api/weather_models/GDPS/predictions/most_recent/ | weather_models/endpoints/test_models_predictions_most_recent_GDPS_[956]_crud_mappings.json      | 200                  | test_models_predictions_most_recent_GDPS_[956]_response.json      | Testing for rare case, where model run time matches prediction time, and previous precip delta is not available. |
