+++ api/app/tests/weather_models/test_machine_learning.py
@@ -2,7 +2,6 @@
 result.
 """
 from datetime import datetime
-from email import parser
 from typing import List
 import json
 import pytest
@@ -10,6 +9,7 @@ from pytest_bdd import scenario, given, then, when, parsers
 from app.db.models import PredictionModel, PredictionModelGridSubset
 from app.weather_models import machine_learning
 from app.tests.weather_models.crud import get_actuals_left_outer_join_with_predictions
+from app.tests.common import str2float
 
 
 @pytest.fixture()
@@ -25,7 +25,9 @@ def test_machine_learning():
     """ BDD Scenario for predictions """
 
 
-@given(parsers.parse("An instance of StationMachineLearning for {coordinate} within {points}"), target_fixture='instance', converters={'coordinate': json.loads, 'points': json.loads})
+@given(parsers.parse("An instance of StationMachineLearning for {coordinate} within {points}"),
+       target_fixture='instance',
+       converters=dict(coordinate=json.loads, points=json.loads))
 def given_an_instance(coordinate: List, points: List):
     """ Bind the data variable """
     return machine_learning.StationMachineLearning(
@@ -44,7 +46,8 @@ def learn(instance: machine_learning.StationMachineLearning):
     instance.learn()
 
 
-@then(parsers.parse('The {model_temp} for {timestamp} results in {bias_adjusted_temp}'), converters={'model_temp': float, 'timestamp': datetime.fromisoformat, 'bias_adjusted_temp': lambda value: None if value == 'None' else float(value)})
+@then(parsers.parse('The model_temp: {model_temp} for {timestamp} results in {bias_adjusted_temp}'),
+      converters=dict(model_temp=float, timestamp=datetime.fromisoformat, bias_adjusted_temp=str2float))
 def assert_temperature(
         instance: machine_learning.StationMachineLearning,
         model_temp: float, timestamp: datetime, bias_adjusted_temp: float):
@@ -53,7 +56,11 @@ def assert_temperature(
     assert result == bias_adjusted_temp
 
 
-@then(parsers.parse('The {model_rh} for {timestamp} results in {bias_adjusted_rh}'), converters={'model_rh': float, 'timestamp': datetime.fromisoformat, 'bias_adjusted_rh': lambda value: None if value == 'None' else float(value)})
+@then(parsers.parse('The model_rh: {model_rh} for {timestamp} results in {bias_adjusted_rh}'),
+      converters=dict(
+          model_rh=float,
+          timestamp=datetime.fromisoformat,
+          bias_adjusted_rh=str2float))
 def assert_rh(instance: machine_learning.StationMachineLearning,
               model_rh: float, timestamp: datetime, bias_adjusted_rh: float):
     """ Assert that the ML algorithm predicts the relative humidity correctly """
